Update deps and fix types, move generated files in to __generated__ folder
