+++ package.json
@@ -12,7 +12,7 @@
   "devDependencies": {
     "@types/js-cookie": "^2.2.6",
     "@types/styled-components": "^5.1.3",
-    "tina-graphql-gateway-cli": "^0.2.27"
+    "tina-graphql-gateway-cli": "^0.2.32"
   },
   "dependencies": {
     "@tailwindcss/ui": "^0.3.0",
@@ -41,7 +41,7 @@
     "styled-components": "^5.1.3",
     "styled-jsx": "^3.2.5",
     "tailwindcss": "^1.4.6",
-    "tina-graphql-gateway": "^0.1.52",
+    "tina-graphql-gateway": "^0.1.53",
     "tinacms": "0.31.0",
     "typescript": "^3.8.3",
     "urql": "^1.9.7",
